🔨 [projects] Move type alias `GenerateProject` to `repository` module
