--- src/cutty/projects/common.py
@@ -1,6 +1,0 @@
-"""Common functionality for projects."""
-from collections.abc import Callable
-from pathlib import Path
-
-
-GenerateProject = Callable[[Path], None]
