+++ tests/unit/projects/test_update.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from cutty.projects.common import GenerateProject
+from cutty.projects.repository import GenerateProject
 from cutty.projects.repository import LATEST_BRANCH
 from cutty.projects.repository import ProjectRepository
 from cutty.projects.repository import UPDATE_BRANCH
